+++ lib/music_theory.dart
@@ -4,5 +4,5 @@ export 'src/note_names.dart';
 
 export 'src/note.dart';
 export 'src/scale.dart';
-export 'src/chord.dart';
+export 'src/chord/chord.dart';
 export 'src/mode.dart';
