+++ lib/src/chord/chord_type.dart
@@ -1,0 +1,14 @@
+enum ChordType {
+  major,
+  minor,
+  augmented,
+  diminished,
+  major7,
+  minor7,
+  dominant7,
+  augmented7,
+  dominant7b5,
+  minorMaj7,
+  halfDiminished,
+  diminished7,
+}
