+++ lib/src/chord/semitone_formula_to_chord_type.dart
@@ -1,0 +1,5 @@
+import 'package:music_theory/src/chord/chord_type.dart';
+
+/// TODO: Implement semitoneListToChordType
+ChordType semitoneFormulaToChordType(List<int> semitonesList) =>
+    ChordType.major;
